# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Open Wallet API.",
        "title": "Open Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List categories",
                "operationId": "v1ListCategories",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "name:asc",
                        "description": "Sort field. \n- Allowed: name, created_at, id",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter expression. \n- Allowed fields \u0026 ops:\n  - color: eq, in\n  - created_at: eq, gt, gte, lt, lte\n  - id: eq, in\n  - name: eq, like, in\n  - user_id: eq, in\n",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "A category name to filter by",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.PaginatedResponse-internal_resources_categories_handlers_ListCategoriesResponseData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a category",
                "operationId": "v1CreateCategory",
                "parameters": [
                    {
                        "description": "Category payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_resources_categories_handlers.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_categories_handlers_CreateCategoryResponseData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{category_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Delete Category By ID",
                "operationId": "v1DeleteCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category deleted"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Update Category By ID",
                "operationId": "v1UpdateCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_resources_categories_handlers.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_categories_handlers_UpdateCategoryResponseData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{period}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List categories with amount per period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List categories with amount per period",
                "operationId": "v1ListCategoryAmountPerPeriod",
                "parameters": [
                    {
                        "type": "string",
                        "description": "period",
                        "name": "period",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter expression. \n- Allowed fields \u0026 ops:\n  - color: eq, in\n  - id: eq, in\n  - name: eq, like, in\n  - period: eq, in\n  - total_amount: eq, gt, gte, lt, lte\n  - user_id: eq, in\n",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "name:asc",
                        "description": "Sort field. \n- Allowed: name, total_amount, period, id",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories with amount per period",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util.PaginatedResponse-internal_resources_categories_handlers_ListCategoryAmountPerPeriodResponseData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_felipe1496_open-wallet_internal_util.PaginatedResponse-internal_resources_categories_handlers_ListCategoriesResponseData": {
            "type": "object",
            "required": [
                "data",
                "query"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_categories_handlers.ListCategoriesResponseData"
                },
                "query": {
                    "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_querybuilder.Metadata"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util.PaginatedResponse-internal_resources_categories_handlers_ListCategoryAmountPerPeriodResponseData": {
            "type": "object",
            "required": [
                "data",
                "query"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_categories_handlers.ListCategoryAmountPerPeriodResponseData"
                },
                "query": {
                    "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_querybuilder.Metadata"
                }
            }
        },
...
        "github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_categories_handlers_CreateCategoryResponseData": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_categories_handlers.CreateCategoryResponseData"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util.ResponseData-internal_resources_categories_handlers_UpdateCategoryResponseData": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_resources_categories_handlers.UpdateCategoryResponseData"
                }
            }
        },
...
        "github_com_felipe1496_open-wallet_internal_util_httputil.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_felipe1496_open-wallet_internal_util_httputil.HTTPErrorData"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util_httputil.HTTPErrorData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_felipe1496_open-wallet_internal_util_querybuilder.Metadata": {
            "type": "object",
            "required": [
                "next_page",
                "page",
                "per_page",
                "total_items",
                "total_pages"
            ],
            "properties": {
                "next_page": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "internal_resources_categories_handlers.CategoryAmountPerPeriodResource": {
            "type": "object",
            "required": [
                "color",
                "id",
                "name",
                "total_amount"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
        "internal_resources_categories_handlers.CategoryResource": {
            "type": "object",
            "required": [
                "color",
                "created_at",
                "id",
                "name",
                "user_id"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_resources_categories_handlers.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "color",
                "name"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_resources_categories_handlers.CreateCategoryResponseData": {
            "type": "object",
            "required": [
                "category"
            ],
            "properties": {
                "category": {
                    "$ref": "#/definitions/internal_resources_categories_handlers.CategoryResource"
                }
            }
        },
        "internal_resources_categories_handlers.ListCategoriesResponseData": {
            "type": "object",
            "required": [
                "categories"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_resources_categories_handlers.CategoryResource"
                    }
                }
            }
        },
        "internal_resources_categories_handlers.ListCategoryAmountPerPeriodResponseData": {
            "type": "object",
            "required": [
                "categories"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_resources_categories_handlers.CategoryAmountPerPeriodResource"
                    }
                }
            }
        },
        "internal_resources_categories_handlers.UpdateCategoryRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                }
            }
        },
        "internal_resources_categories_handlers.UpdateCategoryResponseData": {
            "type": "object",
            "required": [
                "category"
            ],
            "properties": {
                "category": {
                    "$ref": "#/definitions/internal_resources_categories_handlers.CategoryResource"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}